+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.8.0] Q2 2022
+ - added iot edge module support:
+   - bumb to azure-iot-sdk-sys 0.3.0
+   - introduced cargo features to select client type (device, module, edge)
+   - conditional compilation depending on client type
+
 ## [0.7.0] Q2 2022
  - removed all avoidable unwraps()
  - switch to impl Into<String> instead of &str or String in public functions